feat: split content into components #64
