+++ src/App.tsx
@@ -2,7 +2,10 @@ import "./App.css";
 import React, { useState } from "react";
 import Header from "./modules/Header/Header";
 import Footer from "./modules/Footer/Footer";
-import Content from "./modules/Content/Content";
+import ContentTV from "./modules/Content/Tv";
+import WatchOfline from "./modules/Content/WatchOfline";
+import WatchEverywhere from "./modules/Content/WatchEverywhere";
+import KidsProfiles from "./modules/Content/KidsProfiles";
 import Modal from "./modules/Modal/Modal";
 
 export default function App() {
@@ -13,14 +16,21 @@ export default function App() {
   };
 
   return (
-    <div className="wrapper">
-      <Header updateData={updateData} isOpenModal={isOpenModal} />
-      {isOpenModal ? (
-        <Modal updateData={updateData} isOpenModal={isOpenModal} />
-      ) : (
-        <Content />
-      )}
-      <Footer />
-    </div>
+    <>
+      <div className="wrapper">
+        <Header updateData={updateData} isOpenModal={isOpenModal} />
+        {isOpenModal ? (
+          <Modal updateData={updateData} isOpenModal={isOpenModal} />
+        ) : (
+          <>
+            <ContentTV />
+            <WatchOfline />
+            <WatchEverywhere />
+            <KidsProfiles />
+          </>
+        )}
+        <Footer />
+      </div>
+    </>
   );
 }
