+++ src/modules/Content/KidsProfiles.tsx
@@ -1,0 +1,19 @@
+export default function Content() {
+  return (
+      <>
+  <section className="kids-profiles">
+    <div className="kids-profiles__img picture-block"></div>
+    <div className="kids-profiles__content">
+      <h2 className="kids-profiles__content-title">
+        Create profiles for kids.
+      </h2>
+      <h3 className="kids-profiles__content-subtitle">
+        Send kids on adventures with their favorite characters in a space
+        made just for them—free with your membership.
+      </h3>
+    </div>
+  </section>
+  </>
+  );
+};
+
