+++ src/modules/Content/WatchEverywhere.tsx
@@ -1,0 +1,17 @@
+export default function Content() {
+  return (
+      <>
+  <section className="watch-everywhere">
+    <div className="watch-everywhere__content">
+      <h2 className="watch-everywhere__content-title">Watch everywhere.</h2>
+      <h3 className="watch-everywhere__content-subtitle">
+        Stream unlimited movies and TV shows on your phone, tablet, laptop,
+        and TV without paying more.
+      </h3>
+    </div>
+    <div className="watch-everywhere__img picture-block"></div>
+  </section>
+  </>
+  );
+};
+
