+++ src/modules/Content/WatchOfline.tsx
@@ -1,0 +1,18 @@
+export default function Content() {
+  return (
+      <>
+  <section className="watch-offline">
+    <div className="watch-offline__img picture-block"></div>
+    <div className="watch-offline__content">
+      <h2 className="watch-offline__content-title">
+        Download your shows to watch offline.
+      </h2>
+      <h3 className="watch-offline__content-subtitle">
+        Save your favorites easily and always have something to watch.
+      </h3>
+    </div>
+  </section>
+  </>
+  );
+};
+
