+++ src/modules/Header/Header.css
@@ -1,22 +1,29 @@
 .header {
-    display: flex;
-    justify-content: space-between;
-    padding: 30px 60px;
-  }
-  .logo {
-    background-image: url(../../img/Netflix_2015_logo.svg);
-    background-size: cover;
-    width: 170px;
-    height: 45px;
-  }
-  
+  display: flex;
+  justify-content: space-between;
+  padding: 30px 60px;
+}
+.logo {
+  background-image: url(../../img/Netflix_2015_logo.svg);
+  background-size: cover;
+  width: 170px;
+  height: 45px;
+}
+
+.btnSignIn {
+  background-color: #e50914;
+  width: 84px;
+  height: 34px;
+  border-radius: 3px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
+
+@media screen and (max-width: 460px) {
   .btnSignIn {
-    background-color: #e50914;
-    width: 84px;
-    height: 34px;
-    border-radius: 3px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
+    width: 45px;
+    height: 30px;
   }
+}
