+++ src/modules/Modal/Modal.css
@@ -54,3 +54,23 @@
   padding: 10px;
   color: rgb(2, 2, 2);
 }
+
+@media screen and (max-width: 1024px) {
+  .modalBox {
+    width: 55%;
+    height: 85%;
+    border-radius: 5px;
+    background: rgba(0, 0, 0, 0.8);
+  }
+}
+
+
+
+@media screen and (max-width: 756px) {
+  .modalBox {
+    width: 95%;
+    height: 95%;
+    border-radius: 5px;
+    background: rgba(0, 0, 0, 0.8);
+  }
+}
